Added rdfs:comments marked with SES to record issues I noticed on a first pass.
Define equivalent classes for Work or Instance, and Work or Instance or Item.

Added domains and ranges where possible (did not add inferrable domain/ranges).

Changed some literal types to more specific ones where the correct type was obvious.

xsd:string is for sequences of characters that may or may not be in a human language
rdf:PlainLiteral is for  text that may have an associated language, and is equivalent to  (xsd:string or rdf:langString)